requirements.txt
pandas
nump
pickle
sqlalchemy
psycopg2
itertools
os
statsmodels